requirements.txt
# Environment and config
python-dotenv

# HTTP requests
requests

# Database
SQLAlchemy
psycopg2-binary

# Agent framework
crewai  # Assuming you have this locally or from pip

# Dashboard
streamlit
pandas

# Scheduling and orchestration
prefect

# PDF reports
reportlab

# Email support
# smtplib and email.mime are part of the Python Standard Library, so no need to list them

# Slack notifications
slack-sdk

# Optional (handy for .env management)
python-decouple
